[flow][match] Parse sequence expressions as match arg
Summary:
We can and should parse sequence expressions here (which include a comma), even if we want to make it an error in Flow itself later (to avoid confusion with other languages which can provide this as a shorthand for matching on multiple values). This matches Hermes parser implementation. Also factors out this code that is in common between match expressions and statements.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D66530672

fbshipit-source-id: ef2a177a0c42a95282c4523c50d13abbe7abd9a3