[flow] restore normalizer debug
Summary:
This was removed in D54332600 due to dependency cycles. This diff brings this facility back with the caveat that the actual import information is not passed in, which means that the provenance field in symbols will not be accurate. This is still useful for debug printing of types in a shorted form that Debug_js.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D66659185

fbshipit-source-id: b8e71ea14986e0c1b009a31dffd743f60fe30ace