[flow] Slight haste_namespace config tweak and add tests
Summary:
Previously, I reversed the list here because I think the lines in flowconfig are processed in order but they are prepended to the list in reverse order. However, that's my misunderstanding and flowconfig processing will reverse the lines first. This diff removes the reverse here to be consitent with other flowconfig options.

In addition, I also added the compatibility mode that allow common code to import 1-namespace only code (choosing the first namespace).

I also added test to test the namespace resolution behavior and the precedence rule change.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D66710892

fbshipit-source-id: 728f804ddb40032cce62401e8a93a8fb35948fcc