[flow][match] Fix parsing of match statements
Summary:
Unlike match expressions, for match statements we need to rollback when parsing (like we do for parsing arrow expressions). You can take a look at the new test cases for why (previously failing).
Rolling back is expensive, but thankfully it will only happen in the case where we have `match [no newline] (`, which limits the damage, and only in the statement case. (In WWW there are no instances of calls with the callee `match` at the start of an expression statement, that I could find using `flow-runner query-ast -e 'CallExpression[callee.name="match"]'`, so this slow case of having to roll back should not be hit often).

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D66530673

fbshipit-source-id: 72e52db853c3ef19bb87db0a357c1f9744bce11e