[flow][refactor] Preserve more information of computed prop
Summary:
In this diff, instead of always spreading synthesized object from the computed prop, we preserve more information of that kind of the computed prop we are dealing with. Only in `add_computed`, we convert them into the synthesized object to be spread.

This refactor allows us to treat different kinds of computed props differently. For example, for computed props that actually resolve to a literal key, we can avoid the spread indirection, and directly call `add_prop`. In later diffs, we can use this additional information to decide what to do when we encounter the string/any typed computed key, which is currently unsound.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D66966735

fbshipit-source-id: 31489ecef02b7265618532d0b09c068e4d490f55