[flow] Separately handle string/any typed computed property vs other invalid computed properties
Summary:
In this diff, we will detect the kinds of computed property types by inlining the check of `WriteComputedObjPropCheckT`. We do not need another concretization pass, since we are already given the concretized keys here.

For string and any typed keys, we will continue to go through the `ObjectExpressionAcc.ComputedProp.NonLiteralKey` variant, which is unsound, but will be fixed later. For number and other kinds of keys, we will error, and as a result, we can safely ignore them.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D66968683

fbshipit-source-id: 1ee1befcd52d88244061efe6ba533c7dbde8391a