[flow] add general and unsound versions of primitive literal types
Summary:
The `StrT`, `NumT`, `BoolT` and `BigIntT` types encompass two kinds of behaviors with vastly different characteristics.
1. When the `literal` portion of the type is some `Literal` then the type behaves opportunistically as the singleton literal type or the general type `string`, `number`, etc.
2. When the `literal` portion is `Truthy` or `AnyLiteral` then they just behave as the general type.

The first behavior is unsound since it allows obviously unsound assertions like
```
let x = 5;
if (true) x = 42;
x as 5;
```
The second is sound.

The goal of this diff is to disentangle the two behaviors into two new types for each kind of primitive, by introducing a "General" version and an "UNSOUND" version (see changes in type.ml).

In later steps in the natural inference project, the unsound types will be completely removed. With that we will remove the implicit conversion of singleton types to unsound types in subtyping_kit.ml. This will require introducing singleton types in places where now just the unsound version is handled.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D66986461

fbshipit-source-id: 204fe0342b0812a6a15b6f5134193ebf7a9348d1