[flow] Do not break up MaybeT, OptionalT on non-distributive conditional type
Summary:
MaybeT and OptionalT should not be destructed unless we explicitly decide that it should have the distributive behavior:

https://github.com/facebook/flow/blob/8e5ecac527cc76be12eae51a5219106dfbe8c69e/src/typing/flow_js.ml#L6554

https://github.com/facebook/flow/blob/8e5ecac527cc76be12eae51a5219106dfbe8c69e/src/typing/flow_js.ml#L6650-L6651

https://github.com/facebook/flow/blob/8e5ecac527cc76be12eae51a5219106dfbe8c69e/src/typing/flow_js.ml#L6659-L6660

In the main recursive flow_js, we did it correctly only for UnionT:

https://github.com/facebook/flow/blob/8e5ecac527cc76be12eae51a5219106dfbe8c69e/src/typing/flow_js.ml#L1715-L1722

but missed the case for `MaybeT` and `OptionalT`. This diff fixes that.

Changelog: [errors] Previously we incorrectly distribute-over-union for maybe or optional input types of conditional type. (e.g. `(?string) extends string ? true : false` is evaluated to `true | false` instead of `false`). This is now fixed, and code that depends on the bug might have new errors.

Reviewed By: panagosg7

Differential Revision: D67145501

fbshipit-source-id: 573933e56fef794ee29709645d48d4754aa874c5