[flow] reposition type guard before use in refinement
Summary:
When refining with type guards it is possible that we end up using the type guard itself as the result of the refinement. In that case that type needs to be repositioned. If not if the type guard happens to be an opaque type then we risk having the reason associated with it point to a remote file, thus making it invalid.

Changelog: [fix] fixed a bug when using opaque types in type guards in files other than the one where the type guard function was defined.

Reviewed By: SamChou19815

Differential Revision: D67153109

fbshipit-source-id: 2702a2e670fcceef3d817349591db4c04d5373cb