[flow][vscode] Check in oss extension and rewrite in TS
Summary:
Internally, we have an internal-only extension that is much better maintained than the current oss one in https://github.com/flow/flow-for-vscode. Notably, it doesn't support the detailed error rendering and refinement highlighting improvements we have made this year.

I want to correct this situation so that we have a consistent baseline of Flow experience. Since so much drift has already happened and we have done nothing to correct this for almost two years, it's clear that the status-quo won't work and we need to do something different so that it's sustainable to keep the internal and the oss extension in sync.

Therefore, in this diff, I:

1. Check in the code from https://github.com/flow/flow-for-vscode mostly as is with regard to runtime behavior. By moving the code here, we can land changes on both internal and oss extension much more easily.
2. Rewrite in TS (mostly syntactic changes): our internal extension is in TS. The VSCode type definition is in TS. There is way too much work to convert the TS libdef to flow. It will also be easier to directly copy the code from internal to oss extension or vise-versa. It might make us look bad that we are not even dogfooding flow in our own repos, but it's still better than it being a major blocker so that the community gets almost nothing.

Apart from these major decisions, I did some minor cleanups:

1. bumping some dependencies
2. bumping bundled flow
3. removing the prettier formatting in extension (we used to return poorly formatted stuff in hover, so we have to use prettier to re-format the hover results)
4. removing most of the unnecessary middleware stuff.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D67163378

fbshipit-source-id: 72a2c03ad884f0be06936920324ccdcfb4e1df13