Allow union types of string or any in computed properties (#9238)
Summary:
Pre-LTI, this is banned because we cannot make the decision of whether to make it a computed property or named property. However, now with LTI, we can make the decision by inspecting the types, and only make singleton literal types to named property and make everything else computed.

Therefore, in this PR, I have relaxed the restriction. We still do a validation to ensure that are no random types like `{...}` being used as the key. If any of the member of the union has an unsupported type, the property will continue to be ignored. Otherwise, we will support it.

Changelog: [feature] Flow now allows using union of strings as the type of computed property keys.

Pull Request resolved: https://github.com/facebook/flow/pull/9238

Reviewed By: panagosg7

Differential Revision: D67106080

Pulled By: SamChou19815

fbshipit-source-id: da8bf260f6d64dc90ccf0b7112142d49feede264