[flow] Remove `react.ref_as_prop=disabled` override in component_syntax test suite
Summary:
The behavior remains mostly the same. The test is asserting that we are actually ignoring the ref prop in component spread.

Before, we ignore it and subtype the ref prop in jsx against `React.ElementRef<typeof Component>`, which gives back `React.RefSetter<void>` and explains the previous error.

Now with ref-as-prop, the ref prop appears as regular prop and subtype against the entire props object with spreaded ref prop intentionally dropped, so we get the prop-missing error.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D66849734

fbshipit-source-id: 870ed0bc794f76bf7daf15eef45862f9a5db053b