[flow] perform call check on Array [react-rule-unsafe-mutation] errors
Summary:
In
```
component Foo(arr: Array<number>) {
    arr.sort((a: number, b: number) => {
        return 1;
    });
    return null;
}
```
the call to `.sort` is an error, but we should still try to populate some type in the entry for `sort`.

This diff makes it so that right after we raise the EPropNotReadable ReactDeepReadOnly error we perform the call as usual.

Changelog: [errors] Flow will now perform a complete function call check even when [react-rule-unsafe-mutation] errors are raised.

Reviewed By: SamChou19815

Differential Revision: D67320626

fbshipit-source-id: fc2c3df9661f43739c405103f59bd57870c0e2c8