[flow][match] Use synthetic binding for match argument to be able to refine
Summary:
Changelog: [internal]

We want to refine the match argument before we destruct it, however it can be an expression, not just an identifier/member.

We can solve this by creating a synthetic binding. We create one here, using the location added in the previous diff. We initialize it with the value of the argument.

Now in future diffs we will be able to refine this value.

Reviewed By: SamChou19815

Differential Revision: D67261334

fbshipit-source-id: 58a9e3abcaf42695a3b7c977a70a3e99c39b61d8