[flow][match] Simple check for match expression exhaustiveness based on refinements
Summary:
Adds a simple check for match expression exhaustiveness based on the argument being refined to `empty`.

In the future, once we have a more advanced analysis, this could serve as an optimization (if the refinements systems has already refined the argument to `empty` by the end of the `match`, we don't need to run the advanced analysis).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D67261340

fbshipit-source-id: cc1a3dc6577af2f7487f6aaf175613f70dd34942