[flow][errors] Sort speculation branches by score
Summary:
I can't believe that we have never done this... We compute score for each branch of speculation to decide how close the user is to the right type, but we only use it to pick the top few branches to show. This diff also uses it to sort the branches, so that the branch that's the most likely to be useful is shown first. This is important for type argument incompatibiity errors in speculation, since it's really the most likely cause and should be displayed first.

Changelog: [fix] For errors involving unions and intersections, we will now show a list of possible causes in a sorted order: the ones that are mostly likely to be the cause will be shown first.

Reviewed By: panagosg7

Differential Revision: D67803854

fbshipit-source-id: 1c2414c68d31e0c13ec4b8e39fb001c33b0fafc2