[flow][get-def] Use the same heuristic for get-def in require expression
Summary:
This diff applies the same heuristics in the previous diff to require expression. The def loc is recorded in the reason of the literal expression.

Changelog: [ide] Go-to-definition for `require('...')` expression will now jump to the default export of an ESM module if available, or the first export of an ESM module.

Reviewed By: panagosg7

Differential Revision: D67875257

fbshipit-source-id: 3d1a47555d448c3872063ee6d2c349e3c64ebd17