[flow][get-def] Make get-def on jsx-intrinsic jumps to nothing
Summary:
Currently, it jumps to itself due to implementation detail.
It first goes through the useful scope-builder flow, but finds that it has no matching uses, so it ends up being treated as a global and do a reason-based get-def.

This diff explicitly makes it jump to nothing if it's an intrinsic element. This is the same as [TS behavior](https://www.typescriptlang.org/play/?#code/JYWwDg9gTgLgBAJQKYEMDG8BmUIjgcilQ3wG4BYAKCqoB4ATYANzgHoA+CyoA).

Changelog: [ide] Go-to-definition on intrinsic jsx elements will jump to nowhere instead of jump to itself.

Reviewed By: gkz

Differential Revision: D67995311

fbshipit-source-id: ae3b3b8ef216285586bfdce65159e3205f329532