[flow] Make get-def on object patterns in match jumps to nothing for literals, to itself for identifiers
Summary:
Currently, the id case jumps to itself due to implementation detail.
It first goes through the useful scope-builder flow because it eventually hit `this#t_identifier`, but finds that it has no matching uses, so it ends up being treated as a global and do a reason-based get-def.

This diff explicitly makes it jump to itself for named keys, and jump to nothing for all other literal keys.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D67995312

fbshipit-source-id: 54879503d62f41be8ff53f68e5ba232155f2af17