[flow] Principled get-def for specifiers in `export {...} from '...'`
Summary:
In this diff, we will handle get-def for specifiers in `export {...} from '...'` explicitly. Currently, we will use reason because these identifiers don't have definition.

This diff explicitly handles it according to the behavior specified in the changelog below.

Changelog: [ide] We slightly changed the go-to-definition behavior for export from statements:
- For `bar` in `export {foo as bar} from '...'`, we will always jump to itself
- For `foo` in `export {foo} from '...'` or `export {foo as bar} from '...'`, we will always jump to the name at the export, if this statement is well typed.
- For `foo` in `export {foo} from '...'` or `export {foo as bar} from '...'`, we will jump to itself if the statement is not well-typed.

Reviewed By: panagosg7

Differential Revision: D68002309

fbshipit-source-id: eca35686e615eeb87fd32365ecaad7ba6771a198