[flow] Handle get-def on global variables explicitly
Summary:
After all the fixes earlier in the stack that gets rid of all the other get-def that's not actually globals accidentally landing here, we can finally do this. There might still be things that are not covered by tests, but at this point we should just fix forward.

In this diff, we use the def-loc info tracked in Builtins.t to power get-def for globals. With this, most of the remaining `Get_def_request.Type` are for import source only, which can be further cleaned up.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D68004358

fbshipit-source-id: 6f6d4b6215630168d51fd442b47432293999e105