[flow][paste-provider] Setup lsp types, parsers, and formatters
Summary:
This stack of diffs aims to auto import on paster support similar to this feature of TS: https://code.visualstudio.com/updates/v1_96#_paste-with-imports-for-javascript-and-typescript

This diff declares the relevant Flow-custom LSP types first, as well as their parsers and formatters. Their uses will be shown in the next diff.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D68074702

fbshipit-source-id: 43f56b00a8cae7622a1c7aa896a3c3a1d06f71f6