[flow][get-def] Explicitly jump to the same place as `require(...)` for `A` in `const A = require(...)`
Summary:
This diff removes the last use of `Get_def_request.Type` in get_def_process_location. This one is particularly problematic. Currently, it's supposed to work like this:

1. Given `const A = require(...)`, we recognize that it's a special pattern_identifier for toplevel require.
2. We jump to the def_loc tracked in reason of A, which happens to be or intentionally and fragilely design to be at the location of `require(...)`
3. In another iteration in GetDef_js in the loop, we will again do a process location for `require(...)` expression, it's designed to jump to the required module file, where the module info is extracted from reasons.
4. Finally, through this convoluted process, we make it jump to the relevant module.

In previous diffs, we killed all the other pattern_identifier special case within require declarator, and now `const A = require(...)` is the only remaining one, so we can now do it in a much more straightforward way: directly go to step 3, and jump to the required module.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D68170641

fbshipit-source-id: 25edba022cd8d57ee37428116a1361eb5cfb6bfe