[flow][get-def] Remove blanket type-based handling for pattern_identifiers within require call binding
Summary:
This diff gets rid of the blanket special casing behavior for all pattern_identifiers within require declarator. Now we will only do it for the toplevel pattern like `const A = require(...)`. From the tests, it practically only changes the behavior for array pattern, which is probably not common.

Why am I doing this? It's blocking the refactor I'd like to do in the next diff, and this special case isn't creating that much value. This is the behavior for the equivalent esm import code anyways:

```
import * as NS from '...';
const [foo] = NS;
//      ^ jump to itself
```

Changelog: [ide] For identifiers of array patterns bind to a require call, go-to-definition will jump to itself.

Reviewed By: panagosg7

Differential Revision: D68170640

fbshipit-source-id: 20937216d7f0bf23cb09f2be1c678e23e209af65