[flow][get-def] Consistently handle identifiers related to an entire module
Summary:
We currently have some ad-hoc handling for this in following two places:

- `import * as NS from '...'`
- `export * as NS from '...'`

but not `const NS = require('...)`. This diff refactors the logic into a helper and make it consistent. Specifically, the new handling for require during find-ref is important so that we can land D67875259 again. (Checking for find-ref purposes so that we don't try to look through the import, so that if get-def jumps to a name, these names won't be spuriously included. See also D48874845, which includeds additional motivations).

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D68174911

fbshipit-source-id: 771bdfdc328dd792c5ddc18583b9046ce2278852