[flow] Use `File_url.create` and `File_url.parse` for document paste protocol
Summary:
This is a followup of D68132119. The scheme doesn't work for windows paths. This diff replaces it with `File_url.create` and `File_url.parse` pair (which is the underlying uri to file_key conversion code used everywhere else) so that the test can also succeed on windows.

Changelog: [internal]

Differential Revision: D68233875