[flow] Use fragment type from the React module instead of the global type
Summary:
The goal is to give users the ability to override the react libdef with a userland .js.flow file, which unblocks the larger goal of killing the overriding behavior of libdef within libdef), while allowing us to override the type of `React.Node` internally to include fbt.

Currently, this is not possible for one type only: `React$FragmentType`. The global type is referenced during jsx fragment checking, and the global type is defined as `$Exports<'react'>['Fragment']`. Since `$Exports` only works for global declare modules, it forces that the definition of react module to stay in the global libdef.

This diff addresses the issue. In file_sig when we are analyzing explicit and implicit requires, we add `react` to implicit require when we see jsx fragment. Then in places where we currently lookup globals, we now do an implicit named import from `react` to grab the type of `React.Fragment`.

With this diff, we can now remove `React$FragmentType` in future diffs. One fewer dollar type and one fewer weird dependency cycles!

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D67990846

fbshipit-source-id: fbd86989f9ca2c01f5ebddd801916b7c70f3ed9b