[flow] Use `File_url.create` and `File_url.parse` for document paste protocol (#9248)
Summary:
Pull Request resolved: https://github.com/facebook/flow/pull/9248

This is a followup of D68132119. The scheme doesn't work for windows paths. This diff replaces it with `File_url.create` and `File_url.parse` pair (which is the underlying uri to file_key conversion code used everywhere else) so that the test can also succeed on windows.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D68233875

fbshipit-source-id: bf8397be9eb23f172fd1565ed98616e49c18d85f