[flow] type_filter does not rely on unsound literal type coercion
Summary:
Under natural inference, we will ban the implicit coercion of `SingletonStrT` to `StrT_UNSOUND` (same for the rest of the primitive types). Filtering in type_filter.ml, however, relied on this coercion and only handled `StrT_UNSOUND` in various refining functions. To ensure filtering work properly after banning the coercion, this diff handles `SingletonStrT` in the same places where `StrT_UNSOUND` is handled in type_filter.

This change is necessary in D67983681, which adds support for inferred type guards, where we'd like to directly start inferring singleton versions of the various types instead of waiting for natural inference.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D68337708

fbshipit-source-id: b7cb43bfd8f5e8055061d5ac3bb47fed16750c42