[flow] add option to create sound versions of literal types in type_filter
Summary:
Without this change in functions like
```
const fn = (x: 'a'| 'b') => x !== 'a';
```
after D67983681 we'd infer a `StrT_UNSOUND('b')` as the type guard for `x`. This is not ideal as this type is directly lifted to an annotation where these unsound types behave particularly poorly.

So, instead, we introduce a mode where:
* the coercions of `SingletonStrT ~> StrT_UNSOUND`, etc. are disallowed
* Type_filter never produces `StrT_UNSOUND`, etc.

For now this change has no effect, because `with_disallowed_unsound_literal_coercsion` is never called. This starts being used in D67983681.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D68339060

fbshipit-source-id: ebb5ed7e7f808f025b7ce4b7665005ffb1fd2cac