[flow][refactor] Generalize ForcingState
Summary:
In order to kill `ModuleT` as a `Type.t`, we still need the capabilities that Type.t provides: notably, the ability to maintain lazy. Note that wrapping the computation in lazy is not enough, since we need defense again cyclic types.

`ForcingState` provides this ability, but right now it's very coupled with Type.t. This diff generalizes it, which allows a new variant `type module_type = ((Reason.t * TypeTerm.moduletype, TypeTerm.t) result, Reason.t) state` to be added in the next diff.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D68312416

fbshipit-source-id: 991620c9a0a9da09dc4e77e585f50279250bae0c