[flow][refactor] Eliminate generation of `ModuleT` within type sig
Summary:
This diff builds upon previous diffs in the stack to continue purging `ModuleT`. This diff completely eliminates the generation of passing of `ModuleT` within type sig merging completely. Instead, they are replaced with `(Type.moduletype, Type.t (any)) result` or lazy variant of it backed by `ForcingState`.

It starts with making various import-export type sig functions operate on  `(Type.moduletype, Type.t) result` instead of `Type.t`. Due to the big blast radius of the change, a large part of the system also needs to be changed. In addition, we need to carefully maintain the exact same laziness of all the operations. Therefore, we did end up with some weird thunks and forcing functions, aims to have the exact same laziness as before.

This change also allows us to remove a lot of constraints, since we are always certain that we are only dealing with either module or any (encoded in a potentially lazy result type), so we don't need any concretization mechanism for module info.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D68312417

fbshipit-source-id: 724f1a7a698ca74c765df20ccb6dbeb96e1c0482