[flow][cleanup] Kill ModuleT
Summary:
With all the previous diff that removes it's usage, the misbehaving ModuleT can finally be killed. It never belongs to the set with other types, as it's only meaningful to describe the import source and never propagate to anything beyond that. It causes nasty case analysis everywhere, including import/export checking, type normalizer, etc. Now it's finally gone!

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D68382129

fbshipit-source-id: 030dccca7bee2b01b7d65c68e349984c83ecc36e