[flow][refactor] Stop creating ModuleT in get_def_process_location
Summary:
For places where we can directly get either any or module, we can directly use the def_loc from reason. We still need to keep the `ModuleTypeDef` case for the `require('...')` stuff. However, that one doesn't store the ModuleT, so we still manage to completely get rid of all dependency on ModuleT in get-def.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D68381876

fbshipit-source-id: 28a0ac160e60fc324c29dd1bb63e569b48b9caea