[flow][match] Update parsing location of match object pattern properties
Summary:
Previously, the location for object pattern properties was set to the location of the first token. Instead it now contains the entire property. If we require an additional location for type-checking purposes, I will add another one. Thanks to the previous diff, this has no use-facing changes.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D68523163

fbshipit-source-id: a6327705ec65f02d6b6a560292f381737f4605d9