[flow] Create autoimport patches for apply-code-action command
Summary:
Wires up everything from below the stack. Creates the edit commands, converts it from the LSP format to a patch format, and uses the patch printer to print it.

There's an in-place flag so you can print the result to stdout or overwrite the file. Both are tested in the test.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D68646253

fbshipit-source-id: e4e4cf3064caf13f0b0fe8de657832598813bd12