[flow] infer type guards
Summary:
This diff adds initial support for inferring type guards for arrow functions.

For now, type guard inference only supports *arrow functions*, with a *simple expression body*, a *single parameter*, a body the encodes a predicate that is not the "truthy" predicate. So, for example `x => x` is not eligible. The predicate also needs to refine the type of the parameter to be narrower than the input type. So for example, `(x: mixed) => x !== 1` is not eligible, since the `x` is still `mixed` after the predicate `!== 1` is applied.

The inferred type guard is *one-sided*. This simplifies consistency checking significantly, since we don't need to check the inverse direction, and is enough to allow use in Array.filter.

In terms of implementation:
* We extend `record_type_guard_maps` to also consider "candidates" of inferred type guard functions.
* For these candidates we store entries to the `type_guard_consistency_maps`.
* These entries are inspected in type_guard.ml (called from statement.ml) when we need to infer the type guard type.
* When inferring type guards we are disabling "unsound literal coercion" (see D68339060). This is to avoid lifting unsound literal types to type annotation level.
* In `Type_guard.infer_type_guard` under synthesis mode we assume type guard is None. This helps avoid some `ForcedReadOfUnderResolutionTvar` errors and empirically causes fewer errors due to bad hint decomposition (tests/type_guards/inferred.js - test16 is a regression test against this).

Potential extensions:
* allow multiple params
* lift one-sided restriction

Reviewed By: SamChou19815

Differential Revision: D67983681

fbshipit-source-id: a2615207eda1f0ab3559150fadbb1294c836d40f