[flow] check that inferred type guard return expression is boolean
Summary:
For compatibility with declared type guard functions, we want to ensure that an inferred type guard function body always return a boolean expression. This should be mostly guaranteed since we exclude truthy predicates from consideration.

This diff makes this check more direct by checking that the return expression is typed as a boolean.

To do this check we call `Statement.expression` from within type_guard.ml. The dependency requirements for checking the return expression should already be met, since we are already using the expression to infer the type of the guard.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D68730841

fbshipit-source-id: 1bb6a603fac8bfb5919ae0690784d717dbb675d1