[flow] add test showing bad type guard negation behavior
Summary:
Today, in the code
```
function foo(
  isFstParamNum: (value: mixed, extra_value: mixed) => value is number,
  x: number | string,
  y: number | string,
) {
  if (isFstParamNum(x, y)) {
    return;
  }
  x as string; // TODO okay (all LatentP refinements are negated together)
}
```
we generate two "single" `LatentR` refinements Rx and Ry to potentially refine `x` and `y` due to the call to `isFstParamNum`. We need both because whether the call actually refines these variables is determined much later when types are involved.

In the then-branch, these refinements are combined as: And (Rx, Ry).

In the else-case they are negated as: Or (Not Rx, Not Ry)

Ry ends up being a no-op as a refinement so it has zero refining contribution to the end result, and so the `x` will end up being `number|string`.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D68849700

fbshipit-source-id: b564ae00b3ff441b3e973c58c0d3088fc12d35bf