[flow] Make this type after unbinding to be any
Summary:
We do not gain much from the precision here. Consider

```
class Foo {
  foo() { ... }
}
declare const foo: Foo;
const f = foo.foo; // error: method-unbinding
f(); // currently: `global-this ~> Foo error`
```

The second error isn't that necessary, because we already did our job to error at the method-unbinding site.

However, keeping the error makes it harder to fix the larger problem of inconsistent this typing for class methods. With D68961697 but without this diff, we will still get error like `global this is incompatible with implicit class this` even if we explicitly accept the unsoundness of method-unbinding due to CJS export extraction.

To address this, during method-unbinding, I will make the this-type to be any.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D68961698

fbshipit-source-id: e77915e18a03f487a5a1de646a263c8c0500728e