[flow] Align this typing behavior between subtyping and unification
Summary:
The converter function `subtype_this_of_function`, first introduced in D26489429, are called before subtyping. However, it's not done during unification, which causes inconsistency between subtyping and unification and breaks the assumption that `A = B` is equivalent to `A <: B & A :> B`. More importantly, calling it will mitigate the bad behavior exposed in D68902214.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D69025734

fbshipit-source-id: 9f527d95578b9afe7b532e833f7e3f66c77dd501