[flow] add support for "this" type guard
Summary:
This diff adds initial support for this-based type guards, which are only allowed as the return annotation on non-static declare class and interface methods. So the following are allowed:
```
declare class A {
  m(): this is B;
}
interface I {
  m(x: mixed): this is A;
}
```
but these are not
```
declare class B {
  static m(): this is D;
}
type T = (x: mixed): this is A;
function foo(this: mixed): this is A;
class A {
  m(): this is B { return this instanceof B; }
}
```

The main change compared to regular type guards is the addition of a new `LatentP` variant, called `LatentThisP`. This targets `x` in conditional expressions of the form:
```
if (x.m()) {}
```

Due to some issue in name_resolver, `x.f(x)` would cause a crash if we considered `x` as having both a LatentR and a LatentThisR refinement. So we only consider the LatertR one.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D68716825

fbshipit-source-id: 321889c7a110c4cf24d7f5caa1bc99fcc0ce0a5c