[flow] run Env_resolution.resolve_pred_func in synthesis mode
Summary:
This diff reuses the approach of hint computation when computing the types of callees to extract a potential function predicate.

The reason we'd want something like this is to avoid any unwanted side-effects when inferring the type of `e.m` in `e.m()`. For example, when inferring `e.m` in isolation we might end up with new errors like `method-unbinding` or slightly different errors due to calling GetPropT instead of MethodT.

To avoid this we compute the type of the callee in synthesis mode (no node caching) and reset any generated error (like in hint computation).

This diff also tidies up the former `has_new_non_lint_errors` check into `is_lint_only_errorset`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D68934771

fbshipit-source-id: d3704915c95aa06a3f2eda636c8e788c3c8b8e3f