[flow] Enable type sig support for `declare global {...}`
Summary:
This diff connects the type sig support for `declare global {...}` to the rest of the system. Under this diff, and when `experimental.declare_global` is on, importing module with `declare global` will now add globals. However, the local checking support is still not done yet (so spurious errors will appear), which will be done in future diffs.

For now, I only enabled the effect for react module, for the reason described in the comments.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D68721810

fbshipit-source-id: d345cec57ab03a82b9bbfb62edfbbc9ab6ebb3b0